A flag to adjust canonical types to fixed-width typedefs.

+++ test/output_c/include/std_shared_ptr_const_int.h
@@ -53,7 +53,7 @@ MR_C_API const int *MR_C_std_shared_ptr_const_int_Get(const MR_C_std_shared_ptr_
 /// This being zero usually conincides with `MR_C_std_shared_ptr_const_int_Get()` returning null, but is ultimately orthogonal.
 /// Note that in multithreaded environments, the only safe way to use this number is comparing it with zero. Positive values might change by the time you get to use them.
 /// Parameter `_this` can not be null.
-MR_C_API long MR_C_std_shared_ptr_const_int_UseCount(const MR_C_std_shared_ptr_const_int *_this);
+MR_C_API int MR_C_std_shared_ptr_const_int_UseCount(const MR_C_std_shared_ptr_const_int *_this);
 
 /// Create a new instance, taking ownership of an existing pointer.
 /// Parameter `ptr` should point to a single object rather than to an array.
